Avoiding circular import by renaming pandas.py
