+++ pyphi/models/pyphi_pandas.py
